fixes broken link to `app.speakeasy.dev`
In Markdown `[foo.com]()` doesn't link to `foo.com`, it creates a link to the current page, kinda like `.`, so this link that was supposed to point to the app instead pointed to [this non-existent page](https://github.com/speakeasy-api/speakeasy/blob/main) on GitHub. 

Fixed by adding the URL into the `()` as well.