fix: prevent mermaid errors from crashing workflow
This change removes a panic from the mermaid visualization code that
was causing workflow runs to fail when the generated graph was deeper
than two levels.

We are now treating all mermaid errors as non-fatal and instead logging
errors and printing an appropriate message to the user in the action
summary.