fix: TUI tab border styles (#956)
The interactive tabs in the terminal currently try to use unicode box
drawings to render active/inactive rounded tabs. Unfortunately for the
inactive tabs, the active color spills into the active tab styles (see
screenshots). I don't think that it is possible to fix this issue with
the current styles with our glyph borders, so I think personally it's
nicer if we render a hidden border for the inactive tabs

Before:

![CleanShot 2024-09-20 at 14 25
44@2x](https://github.com/user-attachments/assets/b150ebff-3de8-4fc6-829c-cb5a2d99d004)

After:

![CleanShot 2024-09-20 at 14 25
12@2x](https://github.com/user-attachments/assets/5d855bf2-6414-45ff-80ce-9f8cae012597)