fix: better additive method name suggestions (#1044)
https://linear.app/speakeasy/issue/SPE-4404/bug-studio-not-suggesting-additional-fixes-for-newly-added-operations

The actual issue was just a missing `not` in a filter expression. But I
took the opportunity to improve how we get suggestions more broadly by
pre-filtering operations for which we have suggestions. This speeds up
the roundtrip to the LLM (before, we were getting all suggestions then
filtering afterwards)