feat: base quickstart on remote existing target (#999)
This PR adds the ability to bootstrap a new target using a pre-existing
specification hosted on Speakeasy's registry. The CLI will now show the
most recent 5 targets that have been interacted with (based on the CLI
event stream).

If there are any recent target candidates in the workspace, then the
user will be presented with the option to choose from an existing SDK or
create from scratch at the beginning of Quickstart. If there are no
previous generations in a workspace, then the existing "new user" flow
will be shown where the user can enter a filepath to a spec:

![CleanShot 2024-10-07 at 13 46
20@2x](https://github.com/user-attachments/assets/8cc576bc-e0bf-4a56-8ad6-5d5308fb8cae)

Then selecting from the list of recent SDKs:

![CleanShot 2024-10-07 at 13 46
32@2x](https://github.com/user-attachments/assets/6d2b44d3-f93c-4c15-b41f-5e70fb954bea)

(n.b the look and feel of the select list in the 2nd screenshot will
change / look much better once the corresponding
https://github.com/charmbracelet/huh/pull/424 pr has been merged)

The end result is that a new workflow file is created that references
the existing Registry URI for the spec from the previous target

## Next steps

- Try and re-use as much of the prior target's `gen.yaml`