feat: format transform (#1094)
This pull request introduces a new command to format OpenAPI documents
and includes corresponding tests. The most important changes include
adding the `formatCmd` command, implementing the `runFormat` function,
and creating tests to ensure the formatting works correctly.

### New command for formatting OpenAPI documents:

*
[`cmd/openapi/transform.go`](diffhunk://#diff-f618debf6dbd5b8b72bdb3a3772aff1757bd578d0810643b6821fb87c7c6f864R5-R17):
Added the `formatCmd` command to the `transformCmd` command group, which
formats an OpenAPI document to be more human-readable.
[[1]](diffhunk://#diff-f618debf6dbd5b8b72bdb3a3772aff1757bd578d0810643b6821fb87c7c6f864R5-R17)
[[2]](diffhunk://#diff-f618debf6dbd5b8b72bdb3a3772aff1757bd578d0810643b6821fb87c7c6f864R80-R87)
*
[`cmd/openapi/transform.go`](diffhunk://#diff-f618debf6dbd5b8b72bdb3a3772aff1757bd578d0810643b6821fb87c7c6f864R118-R127):
Implemented the `runFormat` function to handle the logic for the new
`formatCmd` command.

### Tests for the new format command:

*
[`internal/transform/format_test.go`](diffhunk://#diff-9511b86f9ed3a81e745007649f568b02e25a674c9c43e729263a857331b45352R1-R53):
Added tests to verify the behavior of the new `formatCmd` command,
ensuring that the formatted OpenAPI document matches the expected
output.