fix(release): fix completion generation on windows
The GitHub action that does the release actually runs on Windows (likely
makes publishing to Choco easier?). The changes to add completions for
Homebrew caused errors during release because I had assumed a *nix like
runtime environment for the GHA `release` job (incorrectly).

This replaces the shell script with a tiny go script that is platform
agnostic.

---

Note: like the other changes, I am unable to test this myself (due to
a private repo being referenced in `go.mod`). In order to test the new
script, you should be able to run `go run scripts/generate.go` (which is
what is done in the `.goreleaser.yaml` file).