docs(configure): Add multiple value details to docs for `tests --rebuild`
Implementation wise, `--rebuild` takes a single string. If it's value is `*` (the default) all tests are rebuilt, but you can specify a list of operationIds to clear only specific ones. The way you do that is by passing a single comma separated list of operationIds like `--rebuild listFoo,deleteBar` etc. 

It was unclear from the docs just how to specify multiple entries, so I added a bit more prose to the help output.