feat: set up interruptible generation in Studio (#1319)
GEN-1145

*WARNING* This PR relies on [openapi-generation
PR#2367](https://github.com/speakeasy-api/openapi-generation/pull/2367)
being merged first and should be merged alongside [speakeasy-registry
PR#4134](https://github.com/speakeasy-api/speakeasy-registry/pull/4134)

This PR sets up the ability to:
- Cancel generation before it is complete
- Receive progress udpates from the generator (new steps, file status
changes, main README content)
- Update the Studio UI based on received updates

---
**1. Cancelability**

To control cancelability, a `CancellableGeneration` field was added to
the `Workflow` struct to hold the *cancellable context* and the *cancel
function*.
When this field is non-nil, `g.GenerateWithCancel` will be used instead
of `g.Generate` (c.f. open-generation PR#2367)

---
**2. Streamability**

To control streamability, a `StreamableGeneneration` field was added to
the `Workflow` struct to hold the `OnProgressUpdate` callback function
that will get
called everytime the generator sends an update:
- if a new step has started then `progressUpdate.Step` will be non-nil
- if a file status has changed then the `progressUpdate.File` will
non-nil
- a `progressUpdate.File.IsMainReadme` boolean is also set to indicate
that the file being modified is the Readme.
If it is `true`, then the `progressUpdate.File.Content` will also be set

---
**3. Studio events**

At the moment, the Studio UI is being notified when:
- the `genSDK` and `compileSDK` steps started
- the main `README` content got received